pydantable-rag/pyproject.toml
 [project]
 name = "pydantable-rag"
 version = "0.1.0"
 description = "FastAPI backend for chatting with a local RAG assistant about pydantable."
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
   "fastapi>=0.110",
   "uvicorn[standard]>=0.27",
   "pydantic>=2.0,<3",
   "python-dotenv>=1.0",
   "sqlite-vec>=0.1.6",
   "numpy>=1.24",
   "httpx>=0.27",
   "ollama>=0.2.1",
   "transformers>=4.41",
   "torch>=2.2",
 ]
 
 [project.optional-dependencies]
 dev = [
   "pytest>=8",
   "pytest-asyncio>=0.23",
   "ruff>=0.8",
 ]
 
 [tool.ruff]
 line-length = 88
 target-version = "py310"
 
 [tool.ruff.lint]
 select = ["E", "F", "I", "UP", "B"]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
